azfunction/requirements.txt
flask
scikit-learn
joblib
pandas
azure-functions


